requirements.txt
asgiref==3.8.1
Django==5.0.6
django-environ==0.11.2
django-perms==2.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
simplejwt==2.0.1
six==1.16.0
sqlparse==0.5.0
typing==3.7.4.3
typing_extensions==4.12.2
